requirements.txt

flask==2.3.3
firebase-admin==6.4.0
numpy==1.26.4
scikit-learn==1.3.0
joblib==1.2.0
cloudinary==1.38.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
matplotlib==3.8.0
torch==2.1.0